Change PPC default to POWER 8
POWER8 is the first officially supported ppc64le platform and remains
the baseline for Fedora. These flags are passed by the Fedora build
system, resulting in both being present during compiler invocation.

When the resulting unit tests are run on a POWER8 machine, they
coredump with "Illegal instruction".

But, vectorscan appears to build and pass its unit tests with power8,
and the limiting factor seems to support for VSX. Which exists on
POWER8, and is extended with _optional_ matrix multiply support on
POWER9.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>